Moved fun commands to cog
